dev: avoid multiple calls to create
If we use createClass with create (to initialize the instance)
we cannot mixing this with the Stack constructor since it will
call the create each time a new Stack is created.

Since we only care about stack having `onHandle` and they
are placeholders they were moved to the constructor instead.

Made Stack work well with non empty instances and wrote a little
utility function for runtime.stack so is more readable and clean